+++ src/appLayerCreator.ts
@@ -10,7 +10,7 @@ import * as fileutil from "./fileutil";
 import logger from "./logger";
 import { Config, Layer, Manifest, Options } from "./types";
 import {getManifestLayerType, getLayerTypeFileEnding, unique} from "./utils";
-import {DockerV2, OCI} from "./MIMETypes";
+import { VERSION } from "./version";
 
 const depLayerPossibles = ["package.json", "package-lock.json", "node_modules"];
 
@@ -156,7 +156,7 @@ async function addDataLayer(
 	config.rootfs.diff_ids.push("sha256:" + dhash);
 	config.history.push({
 		created: options.setTimeStamp || new Date().toISOString(),
-		created_by: "doqr",
+		created_by: `doqr:${VERSION}`,
 		comment: comment,
 	});
 }